+++ app/src/main/java/com/wellbees/popularmovies/ui/MovieDetailsFragment.kt
@@ -1,27 +1,27 @@
-package com.wellbees.popularmovies
+package com.wellbees.popularmovies.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.wellbees.popularmovies.databinding.FragmentPersonDetailsBinding
+import com.wellbees.popularmovies.databinding.FragmentMovieDetailsBinding
 
+class MovieDetailsFragment : Fragment() {
 
-class PersonDetailsFragment : Fragment() {
-
-    private lateinit var binding: FragmentPersonDetailsBinding
+    private lateinit var binding: FragmentMovieDetailsBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentPersonDetailsBinding.inflate(inflater)
+
+        binding = FragmentMovieDetailsBinding.inflate(inflater)
+
 
         return binding.root
     }
