+++ app/src/main/java/com/wellbees/popularmovies/ui/PersonDetailsFragment.kt
@@ -1,27 +1,27 @@
-package com.wellbees.popularmovies
+package com.wellbees.popularmovies.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.wellbees.popularmovies.databinding.FragmentMovieDetailsBinding
+import com.wellbees.popularmovies.databinding.FragmentPersonDetailsBinding
 
-class MovieDetailsFragment : Fragment() {
 
-    private lateinit var binding: FragmentMovieDetailsBinding
+class PersonDetailsFragment : Fragment() {
+
+    private lateinit var binding: FragmentPersonDetailsBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-
-        binding = FragmentMovieDetailsBinding.inflate(inflater)
-
+        binding = FragmentPersonDetailsBinding.inflate(inflater)
 
         return binding.root
     }
